requirements.txt
pip
Flask
cassandra-driver
Flask-WTF
requests